+++ utility/cachelist.h
@@ -15,26 +15,26 @@ class CacheNode
 {
 public:
 
-	CacheNode(int maxsize);
+	CacheNode(size_t maxsize);
 	~CacheNode();
 	/************************************************************************/
 	/* push  对应的内容。返回起始地址                                            */
 	/************************************************************************/
-	char* push(void* data,int len);
+	char* push(void* data,size_t len);
 
 	/************************************************************************/
 	/* 返回对应长度的内容，返回实际的长度与指针位置                              */
 	/************************************************************************/
-	void* getdata(int& length);
+	void* getdata(size_t& length);
 
 	/************************************************************************/
-	/* 弹出对应长度的内容，在getdata后调用                                      */
+	/* 弹出对应长度的内容，在getdata后调用 length 需要的长度，return实际的长度                                     */
 	/************************************************************************/
-	int pop(int length);
+	size_t pop(size_t length);
 	//剩下可以放数据的长度
-	int avablesize();
+	size_t avablesize();
 	//剩下未pop的数据长度
-	int size();
+	size_t size();
 
 private:
 
@@ -48,9 +48,9 @@ public:
 	bool isdeleted();
 private:
 	char* m_buf;
-	int m_maxSize;
+	size_t m_maxSize;
 	char* m_cur;
-	int m_length;
+	size_t m_length;
 	bool m_deleteAll;
 };
 
@@ -60,14 +60,14 @@ public:
 	CCacheList();
 	~CCacheList();
 	
-	void* get_data(int& length);
-	void* push_back(void* data, int length);
-	int pop_front(int maxlength);
+	void* get_data(size_t& length);
+	void* push_back(void* data, size_t length);
+	size_t pop_front(size_t maxlength);
 	int is_empty();
 private:
 	 
 	std::list<CacheNode*> m_nodeList;
-	int m_nodeSize;
+	size_t m_nodeSize;
 
 };
